+++ test/src/test/java/org/hiero/mirror/test/e2e/acceptance/client/TokenClient.java
@@ -71,6 +71,16 @@ public class TokenClient extends AbstractNetworkClient {
 
     @Override
     public void clean() {
+        if (Boolean.parseBoolean(System.getenv("HIERO_MIRROR_TEST_ACCEPTANCE_SKIP_CLEANUP_ENTITIES"))) {
+            // In CI we don't want to cleanup as the entities are needed in the k6 test in the next step.
+            log.warn("Acceptance tests running in CI -> skip cleanup.");
+            for (var tokenName : tokenMap.keySet()) {
+                log.info("Skipping cleanup of token [" + tokenName.getSymbol() + "] at address "
+                        + tokenMap.get(tokenName).tokenId().toEvmAddress());
+            }
+            return;
+        }
+
         var admin = sdkClient.getExpandedOperatorAccountId();
         log.info("Deleting {} tokens and dissociating {} token relationships", tokenIds.size(), associations.size());
         deleteAll(tokenIds, tokenId -> delete(admin, tokenId));
